+++ src/Http/Presenters/DebugPresenter.php
@@ -4,14 +4,17 @@ namespace Fluxoft\Rebar\Http\Presenters;
 use Fluxoft\Rebar\Http\Response;
 
 class DebugPresenter implements PresenterInterface {
-	public function Render(Response $response, array $data): void {
+	/** @inheritDoc */
+	public function Format(array $data): array {
 		$body  = "*** The page's data set: ***\n\n";
 		$body .= $this->renderData($data);
 		$body .= "\n****************************\n";
 
-		$response->AddHeader('Content-type', 'text/plain');
-		$response->Body = $body;
-		$response->Send();
+		return [
+			'body'    => $body,
+			'status'  => 200,
+			'headers' => ['Content-type' => 'text/plain']
+		];
 	}
 
 	private $tab = "	";