+++ src/Http/Presenters/TwigPresenter.php
@@ -4,6 +4,9 @@ namespace Fluxoft\Rebar\Http\Presenters;
 use Fluxoft\Rebar\Exceptions\PropertyNotFoundException;
 use Fluxoft\Rebar\Http\Response;
 use Twig\Environment;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Error\SyntaxError;
 
 /**
  * Class Twig
@@ -12,31 +15,56 @@ use Twig\Environment;
  * @property string $Template
  */
 class TwigPresenter implements PresenterInterface {
+	/**
+	 * TwigPresenter constructor.
+	 * @param Environment $twig The Twig object to use for rendering.
+	 * @param string $template The default template file to use.
+	 */
 	public function __construct(
 		protected Environment $twig,
-		protected string $template = '/default.html.twig',
-		protected string $layout = ''
+		protected string $template = '/default.html.twig'
 	) {}
 
-	public function Render(Response $response, array $data): void {
-		if (!empty($this->layout)) {
-			$data['pageTemplate'] = $this->template;
-			$template             = $this->layout;
-		} else {
-			$template = $this->template;
-		}
-		$output = $this->twig->render($template, $data);
+	/**
+	 * Transform the given data into the desired output format.
+	 *
+	 * @param array $data The data to be rendered in the template.
+	 * @return array{body: string, status: int, headers: array<string, string>}
+	 */
+	public function Format(array $data): array {
+		try {
+			$output = $this->twig->render($this->template, $data);
 
-		$response->Body = $output;
-		$response->Send();
+			return [
+				'body' => $output,
+				'status' => 200,
+				'headers' => ['Content-Type' => 'text/html']
+			];
+		} catch (LoaderError $e) {
+			return [
+				'body' => 'Template not found: ' . $this->template,
+				'status' => 404,
+				'headers' => ['Content-Type' => 'text/plain']
+			];
+		} catch (SyntaxError $e) {
+			return [
+				'body' => 'Template syntax error.',
+				'status' => 500,
+				'headers' => ['Content-Type' => 'text/plain']
+			];
+		} catch (RuntimeError $e) {
+			return [
+				'body' => 'Template runtime error.',
+				'status' => 500,
+				'headers' => ['Content-Type' => 'text/plain']
+			];
+		}
 	}
+
 	public function __set($var, $val) {
 		switch ($var) {
 			case 'Template': // @codeCoverageIgnore
 				$this->template = $val;
-				break;
-			case 'Layout': // @codeCoverageIgnore
-				$this->layout = $val;
 				break;
 			default:
 				throw new PropertyNotFoundException(sprintf(
@@ -49,8 +77,6 @@ class TwigPresenter implements PresenterInterface {
 		switch ($var) {
 			case 'Template': // @codeCoverageIgnore
 				return $this->template;
-			case 'Layout': // @codeCoverageIgnore
-				return $this->layout;
 			default:
 				throw new PropertyNotFoundException(sprintf(
 					'The property %s does not exist.',